+++ examples/py/get-started-demo.py
@@ -46,7 +46,7 @@ all_points = npPointsCraft.detect(img, targetBoxes,[5,2,0])
 zones = convertCvZonesRGBtoBGR([getCvZoneRGB(img, reshapePoints(rect, 1)) for rect in all_points])
 
 # predict zones attributes
-regionIds, stateIds, countLines = optionsDetector.predict(zones)
+regionIds, countLines = optionsDetector.predict(zones)
 regionNames = optionsDetector.getRegionLabels(regionIds)
 
 # find text with postprocessing by standart